03-agent-samples/pyproject.toml
[project]
name = "agent-samples"
version = "0.1.0"
description = "Sample Python scripts for using Azure AI Foundry Agents with APIM-proxied LLM endpoints"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-projects>=2.0.0b2",
    "azure-identity>=1.26.0b1",
    "python-dotenv>=1.2.1",
]
